Restructure files to avoid circular dependencies for Bazel
