+++ compiler/src/iree/compiler/Codegen/LLVMGPU/Passes.h
@@ -177,15 +177,6 @@ void registerCodegenLLVMGPUPasses();
 
 std::unique_ptr<OperationPass<ModuleOp>> createTestLLVMGPULegalizePass();
 
-//------------------------------------------------------------------------------
-// Public functions
-//------------------------------------------------------------------------------
-
-// Loop prefetching transformation. Returns the new loop on success or failure
-// when the `forOp` is not supported.
-FailureOr<scf::ForOp> prefetchSharedMemoryCopy(RewriterBase &rewriter,
-                                               scf::ForOp forOp);
-
 } // namespace mlir::iree_compiler
 
 #endif // IREE_COMPILER_CODEGEN_LLVMGPU_PASSES_H_
