+++ compiler/src/iree/compiler/Codegen/LLVMGPU/Utils/BUILD.bazel
@@ -20,6 +20,7 @@ iree_compiler_cc_library(
         "AMDGPUDistributionPatterns.cpp",
         "LLVMGPULayoutAnalysisAndDistribution.cpp",
         "LLVMGPUUtils.cpp",
+        "PrefetchSharedMemoryCopy.cpp",
     ],
     hdrs = [
         "LLVMGPUUtils.h",
@@ -29,6 +30,7 @@ iree_compiler_cc_library(
         "//compiler/src/iree/compiler/Codegen/Common/GPU:CommonGPUPasses",
         "//compiler/src/iree/compiler/Codegen/Transforms",
         "//compiler/src/iree/compiler/Codegen/Utils",
+        "//compiler/src/iree/compiler/Dialect/HAL/IR",
         "//llvm-external-projects/iree-dialects:IREEVectorExtDialect",
         "@llvm-project//llvm:Support",
         "@llvm-project//mlir:AMDGPUDialect",
@@ -41,6 +43,9 @@ iree_compiler_cc_library(
         "@llvm-project//mlir:MathDialect",
         "@llvm-project//mlir:MemRefDialect",
         "@llvm-project//mlir:NVGPUDialect",
+        "@llvm-project//mlir:SCFDialect",
+        "@llvm-project//mlir:SCFTransforms",
+        "@llvm-project//mlir:Support",
         "@llvm-project//mlir:VectorDialect",
     ],
 )
